% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{dataCollision}
\alias{dataCollision}
\title{Ungulate-Vehicle Collisions in Nine French Departements}
\format{A list with three elements named \code{RedDeer},
\code{RoeDeer} and \code{WildBoar}, each element being itself a
list containing the following elements:

\describe{

\item{X}{a data.frame containing the variables describing the
management units, and supposed to have an effect on the collisions:
(i) \code{forFrag}: forest fragmentation, (ii) \code{urbFrag}:
urban fragmentation, (iii) \code{locr}: density of local roads,
(iv) \code{regr}: density of regional roads, (v) \code{natr}:
density of national roads, (vi) \code{motr}: density of motorways,
(vii) \code{elev600} percentage of the management unit covered by
areas with elevation lower than 600 m asl, (viii)
\code{elev600_1500} percentage of the management unit covered by
areas with elevation comprised between 600 m and 1500 m asl, (ix)
\code{elev600} percentage of the management unit covered by areas
with elevation greater than 1500 m asl, (x) \code{sinus}: average
sinuosity of the roads in the management unit, (xi) \code{hunt}:
the hunting bag, (xii) \code{Agriculture,Open,Urban,Forest}:
proportion of the management unit covered by these habitat types.}

\item{coll}{the number of collisions recorded in each management
unit}

\item{Y}{the number of years during which collisions were recorded
in each management unit}

\item{Area}{the Area of each management unit}

\item{departement}{the departement corresponding to each management unit}
}}
\usage{
dataCollision
}
\description{
Ungulate-Vehicle Collisions in Nine French Departements
}
\details{
Dataset describing the number of Ungulate-Vehicle
collisions in the management units of Nine French Departements.
}
\references{
Saint-Andrieux, C., Calenge, C. and Bonenfant, C. in
prep. Comparison of ecological, biological and anthropogenic causes
of vehicle-wildlife collisions among three large mammalian species
}
\keyword{datasets}

