% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{finalModel}
\alias{finalModel}
\title{Bayesian Model To Identify Factors Affecting Wildlife-Vehicle Collisions}
\usage{
finalModel(X, vectorFinalVariables, collisions, nYear, Area, departement)
}
\arguments{
\item{X}{a data.frame containing the numeric variables supposed to
have an effect on the wildlife-vehicle collisions (columns) for
spatial unit (rows).}

\item{vectorFinalVariables}{a vector of character strings
containing the names of the final variables used in the
combination.}

\item{collisions}{an integer vector with length equal to
\code{nrow(X)} defining the number of wildlife-ungulate collisions
in each studied spatial unit.}

\item{nYear}{an integer vector with length equal to \code{nrow(X)}
defining the number of Years during which wildlife-ungulate
collisions have been recorded in each studied spatial unit.}

\item{Area}{a numeric vector with length equal to \code{nrow(X)}
defining the area of each studied spatial unit.}

\item{departement}{a character vector with length equal to
\code{nrow(X)} defining the department of each studied spatial unit.}
}
\value{
a list with all elements required for the fit of the model
with JAGS, that is: (i) \code{data4jags}: the list of the data
required by the model, to be passed to the argument \code{data} of
the function \code{jags.model} of the package \code{rjags}, (ii)
\code{ini}: list of starting values for the parameters, to be
passed to the argument \code{init} of the function
\code{jags.model}, (iii) \code{modelstring}: a character string
containing the model fit by JAGS, (iv) \code{coefnames}: vector of
character strings containing the names of the coefficients of
interest in the model, to be passed to the argument
\code{variable.names} of the function \code{coda.samples} of the
package \code{rjags}.
}
\description{
Bayesian Model To Identify Factors Affecting Wildlife-Vehicle Collisions
}
\details{
\code{finalModel} fits the final Bayesian model used to
predict the number of collisions between ungulates and vehicle as a
function of a linear combination of a set of environmental
variables.
}
\examples{
## The data used for the fit
data("dataCollision")

## Consider the Kuo-Mallik model fit (see ?prepareFit to see how
## this model was fitted)
data("modelRedDeer")

## we consider the probability associated to each model
probabilityKM(modelRedDeer)

## The best model seems to be the model including forFrag and hunt

## Prepare the data
X <- dataCollision$RedDeer$X

## We scale the variables to improve mixing
X <- as.data.frame(scale(X))

## Prepare the final model, including only the variables forFrag and hunt
pfm <- finalModel(X, c("forFrag","hunt"), dataCollision$RedDeer$coll,
                  dataCollision$RedDeer$Y, dataCollision$RedDeer$Area,
                  dataCollision$RedDeer$departement)

\dontrun{
## WARNING: long execution (about 10 min)!!
## the results are stored in the dataset "finalModelRedDeer"
## But if you want to try it, the data are now ready for the fit:

mo <- jags.model(textConnection(pfm$modelstring), ini=pfm$ini, data=pfm$data4jags)
update(mo, n.iter=1000)
finalModelRedDeer <- coda.samples(mo, variable.names = pfm$coefnames,
                                  n.iter = 500000, thin=100)

}
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link[rjags]{jags.model}},
\code{\link[rjags]{coda.samples}}, \code{\link{prepareFit}}
}

