% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{finalModelRedDeer}
\alias{finalModelRedDeer}
\title{Final Model of the Factors Affecting Ungulate-Vehicle Collisions}
\format{An object of class \code{mcmc.list}}
\usage{
finalModelRedDeer
}
\description{
Final Model of the Factors Affecting Ungulate-Vehicle Collisions
}
\details{
Dataset containing the results of the MCMC fit of the
final model predicting the Red Deer-Vehicle collisions as a
function of the forest fragmentation and hunting bag. The fit was
performed using the package rjags. See the examples of the function
\code{finalModel} to see how this dataset was generated.
}
\seealso{
\code{\link{finalModel}}
}
\keyword{datasets}

