% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{modelRedDeer}
\alias{modelRedDeer}
\title{Kuo and Mallik Model of the Factors Affecting Ungulate-Vehicle Collisions}
\format{An object of class \code{mcmc.list}}
\usage{
modelRedDeer
}
\description{
Kuo and Mallik Model of the Factors Affecting Ungulate-Vehicle Collisions
}
\details{
Dataset containing the results of the MCMC fit of the
model used to identify the variables affecting Red Deer-Vehicle
collisions (as described by Kuo and Mallik, 1998). The fit was
performed using the package rjags. See the examples of the function
\code{prepareFit} to see how this dataset was generated.
}
\references{
Kuo, L. and Mallik, B. 1998.Variable selection for
regression models. Sankhya: The Indian Journal of Statistics,
Series B, 60, 65-81.
}
\seealso{
\code{\link{prepareFit}}
}
\keyword{datasets}

