% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{probabilityKM}
\alias{probabilityKM}
\title{Probability of Variables and Models Calculated with a Kuo-Mallik Approach}
\usage{
probabilityKM(x)
}
\arguments{
\item{x}{an object of class \code{mcmc.list}, returned by the
function \code{coda.samples} from the package \code{rjags}.  The
model string used to fit the model should have been generated by
the function \code{prepareFit}.}
}
\value{
a list with two elements containing: (i) \code{variables}:
a data.frame containing the probability of inclusion of each
variable in the model, (ii) \code{models}: a data.frame containing
the probability that each possible model (i.e. a given combination
of the variables used in the model) is true. In both cases, the
variables and models are sorted by decreasing value of probability.
}
\description{
Probability of Variables and Models Calculated with a Kuo-Mallik Approach
}
\details{
\code{probabilityKM} calculates the probability of
inclusion of the variables, as well as the probability that each
possible model is true, using a Kuo-Mallik approach. The model used
should have been fitted with the package rjags, using the model
string returned by the function \code{prepareFit}.
}
\examples{
data("modelRedDeer")
probabilityKM(modelRedDeer)
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{prepareFit}}, \code{\link[rjags]{coda.samples}}
}

