% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{package}
\name{ungulateCollisions}
\alias{ungulateCollisions}
\alias{ungulateCollisions-package}
\title{ungulateCollisions: Identifying the Factors Affecting Wildlife Ungulate Collisions}
\description{
The package ungulateCollisions provides the dataset as
well as the R code used by Saint Andrieux et al. (in prep.) to
identify the environmental factors affecting the wildlife Ungulate
collisions. This is done using a Bayesian model of the type
described by Kuo and Mallik (1998).
}
\references{
Kuo, L. and Mallik, B. 1998.Variable selection for
regression models. Sankhya: The Indian Journal of Statistics,
Series B, 60, 65-81.

Saint Andrieux, C., Calenge, C. and Bonenfant, C. in
prep. Comparison of ecological, biological and anthropogenic causes
of vehicle-wildlife collisions among three large mammalian species.
}

